Updated drivers for VEML6075 and produced test code to print UV readings to serial console.
