+++ test/serial_print_test.c
@@ -7,55 +7,19 @@
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <unistd.h>
+#include <wiringPi.h>
 
-#include "VEML6075v2.h"
+#include "VEML6075.h"
 
-// I2C Linux device handle
-int g_i2cFid;
-
-// open the Linux device
-void i2cOpen()
-{
-	g_i2cFid = open("/dev/i2c-1", O_RDWR);
-if (g_i2cFid < 0) {
-perror("i2cOpen");
-exit(1);
-	}
-}
-
-// close the Linux device
-void i2cClose()
-{
-close(g_i2cFid);
-}
-
-// set the I2C slave address for all subsequent I2C device transfers
-void i2cSetAddress(int address)
-{
-if (ioctl(g_i2cFid, I2C_SLAVE, address) < 0) {
-perror("i2cSetAddress");
-exit(1);
-	}
-}
-
-
+static int intial_setup = wiringPiSetup();
+static int fd = wiringPiI2CSetup(VEML6075_DEVID);
 
 int main (){
-  i2cOpen();
-
   fprintf("VEML6075 Simple Test \n");
-  i2cSetAddress(VEML6075_ADDR);
 
-  VEML6075 uv = VEML6075();
-
-  if (! uv.begin()) {
-    fprintf("Failed to communicate with VEML6075 sensor \n");
-  }
-  fprintf("Found VEML6075 sensor \n");
+	float UV_calc = UV_sensor::readUVI();
 
   fprintf("UV Index reading: ");
   fprintf(uv.readUVI());
-  delay(1000);
-
-  i2cClose();
+  wait(1);
 }
